Keep the flux full scale an int so the assembler can shift it
ff_maxv returned a float where soccfg['gens'][n]['maxv'] had been an
int, and the tProc assembler bit-shifts the gain field into the
instruction word: TypeError on the first flux pulse of any program
that parks.

+++ WorkingProjects/TLS_Spectroscopy/Client_modules/Helpers/PulseFunctions.py
@@ -510,11 +510,11 @@ def create_qubit_pulse(prog: AcquireProgram, freq: float) -> float:
 
 def ff_maxv(prog, scaled=False):
     gen = prog.soccfg['gens'][int(prog.cfg["ff_ch"])]
-    maxv = float(gen['maxv'])
+    maxv = int(gen['maxv'])
     if not scaled:
         return maxv
     try:
-        return maxv * float(gen['maxv_scale'])
+        return int(maxv * float(gen['maxv_scale']))
     except (KeyError, TypeError, IndexError):
         return maxv
